# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexora Field AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-orders/{id}/checkin:
    get:
      operationId: getCheckin
      tags: [checkin]
      summary: Get checkin record for service order
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinRecord"
    post:
      operationId: doCheckin
      tags: [checkin]
      summary: Perform check-in
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinRecord"
  /service-orders/{id}/checkout:
    post:
      operationId: doCheckout
      tags: [checkin]
      summary: Perform check-out
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckinRecord"

  # PLANS & SUBSCRIPTIONS
...
  /ai/classify/{id}:
    post:
      operationId: classifyOrder
      tags: [ai]
      summary: AI classification of service order
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiClassification"
  /ai/pricing:
    post:
      operationId: getAiPricing
      tags: [ai]
      summary: AI pricing suggestion
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiPricingInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiPricingResult"
  /ai/match/{orderId}:
    get:
      operationId: getAiMatchV2
      tags: [ai]
      summary: AI-powered technician matching
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechnicianMatch"

  # DISPATCH
  /dispatch/{orderId}:
    post:
      operationId: triggerDispatch
      tags: [dispatch]
      summary: Trigger auto-dispatch for a service order
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchResult"
  /dispatch/{orderId}/accept:
    post:
      operationId: acceptDispatch
      tags: [dispatch]
      summary: Accept a dispatch invite
      security: [{ bearerAuth: [] }]
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
components:
  schemas:
...
    Technician:
      type: object
      required: [id, userId, name, email, city, state, specialties, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        cpf:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        photoUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        specialties:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        totalServices:
          type: integer
        totalEarnings:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
...
    TechnicianMatch:
      type: object
      required: [technician, score, reasoning]
      properties:
        technician:
          $ref: "#/components/schemas/Technician"
        score:
          type: number
          description: Match score 0-100
        reasoning:
          type: string

    # COMPANIES
...
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean

    # NOTIFICATIONS
...
    CheckinRecord:
      type: object
      required: [id, serviceOrderId, technicianId]
      properties:
        id:
          type: integer
        serviceOrderId:
          type: integer
        technicianId:
          type: integer
        checkinAt:
          type: ["string", "null"]
          format: date-time
        checkinLatitude:
          type: ["number", "null"]
        checkinLongitude:
          type: ["number", "null"]
        checkinPhotoUrl:
          type: ["string", "null"]
        checkoutAt:
          type: ["string", "null"]
          format: date-time
        checkoutLatitude:
          type: ["number", "null"]
        checkoutLongitude:
          type: ["number", "null"]
        checkoutPhotoUrl:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    CheckinInput:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
        longitude:
          type: number
        photoUrl:
          type: string
    CheckoutInput:
      type: object
      required: [latitude, longitude]
      properties:
        latitude:
          type: number
        longitude:
          type: number
        photoUrl:
          type: string
        notes:
          type: string

    # PLANS & BILLING
...
    AiClassification:
      type: object
      required: [orderId, category, complexity]
      properties:
        orderId:
          type: integer
        category:
          type: string
        complexity:
          type: string
          enum: [baixa, media, alta]
        estimatedHours:
          type: ["number", "null"]
        reasoning:
          type: ["string", "null"]
        aiGenerated:
          type: boolean
    AiPricingInput:
      type: object
      required: [title, description, category, city, state]
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        city:
          type: string
        state:
          type: string
        complexity:
          type: string
    AiPricingResult:
      type: object
      required: [minValue, idealValue, premiumValue]
      properties:
        minValue:
          type: number
        idealValue:
          type: number
        premiumValue:
          type: number
        reasoning:
          type: ["string", "null"]
        aiGenerated:
          type: boolean

    # DISPATCH
...
    DispatchResult:
      type: object
      required: [orderId, dispatched, total]
      properties:
        orderId:
          type: integer
        dispatched:
          type: array
          items:
            type: object
        total:
          type: integer