# lib/api-zod/src/generated/types/checkinRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora Field AI API
 * OpenAPI spec version: 0.1.0
 */

export interface CheckinRecord {
  id: number;
  serviceOrderId: number;
  technicianId: number;
  /** @nullable */
  checkinAt?: Date | null;
  /** @nullable */
  checkinLatitude?: number | null;
  /** @nullable */
  checkinLongitude?: number | null;
  /** @nullable */
  checkinPhotoUrl?: string | null;
  /** @nullable */
  checkoutAt?: Date | null;
  /** @nullable */
  checkoutLatitude?: number | null;
  /** @nullable */
  checkoutLongitude?: number | null;
  /** @nullable */
  checkoutPhotoUrl?: string | null;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  notes?: string | null;
}