# lib/api-zod/src/generated/types/dispatchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nexora Field AI API
 * OpenAPI spec version: 0.1.0
 */
import type { DispatchResultDispatchedItem } from './dispatchResultDispatchedItem';

export interface DispatchResult {
  orderId: number;
  dispatched: DispatchResultDispatchedItem[];
  total: number;
}